# backend/docs/swagger.yaml
basePath: /
definitions:
  auth.AuthResponse:
    properties:
      token:
        type: string
    type: object
  auth.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  auth.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  auth.RegisterRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  rooms.CreateUserRoomRequest:
    properties:
      id:
        type: string
      room_guid:
        type: string
      roomGuid:
        type: string
    type: object
  rooms.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  rooms.Room:
    properties:
      id:
        type: string
      name:
        type: string
      roomUrn:
        type: string
    type: object
  rooms.UserRoom:
    properties:
      assigned_at:
        type: string
      created_at:
        type: string
      id:
        type: string
      room:
        type: string
      room_guid:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Smart Home Hub Backend API
  title: SmartHome Hub API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      - application/x-www-form-urlencoded
      description: Authenticate user and return JWT token
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      - description: Username
        in: formData
        name: username
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Login user
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Create user account
      parameters:
      - description: Register payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Register user
      tags:
      - auth
  /api/rooms:
    get:
      description: Return rooms added to the authenticated user's room list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/rooms.UserRoom'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rooms.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rooms.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List user rooms
      tags:
      - rooms
    post:
      consumes:
      - application/json
      description: Add a predefined room to the authenticated user's room list
      parameters:
      - description: Room to add
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rooms.CreateUserRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/rooms.UserRoom'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rooms.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rooms.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/rooms.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rooms.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add user room
      tags:
      - rooms
  /api/rooms/{id}:
    delete:
      description: Remove a room from the authenticated user's room list
      parameters:
      - description: User room id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rooms.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rooms.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rooms.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user room
      tags:
      - rooms
  /api/rooms/catalog:
    get:
      description: Return the predefined room catalog a user can choose from
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/rooms.Room'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rooms.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rooms.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List predefined rooms
      tags:
      - rooms
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"